update-jira-usernames/requirements.txt
tqdm==4.45.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
pyzmq==19.0.0
requests==2.23.0
